+++ cv_personal_page/cv_personal_page/settings/base.py
@@ -108,7 +108,8 @@ USE_I18N = True
 USE_TZ = True
 
 STATIC_URL = "/static/"
-STATICFILES_DIRS = (BASE_DIR / "static/",)
+STATICFILES_DIRS = (BASE_DIR / "static",)
+STATIC_ROOT = BASE_DIR / "staticfiles"
 
 
 LOCALE_PATHS = (BASE_DIR / "locale/",)